added,CRUD'ed comments

+++ app/Http/Controllers/ArticleController.php
@@ -54,7 +54,8 @@ class ArticleController extends Controller
    */
   public function show(Article $article)
   {
-    return view("articles/show", compact("article"));
+
+    return view('articles/show', compact('article'));
     // return response()->json($article, 200);
   }
 
@@ -63,7 +64,6 @@ class ArticleController extends Controller
    */
   public function edit(Article $article)
   {
-    $this->authorize('update', [self::class, $article]);
     return view('articles/edit', compact('article'));
     // return response()->json($article, 200);
   }
@@ -73,6 +73,7 @@ class ArticleController extends Controller
    */
   public function update(Request $request, Article $article)
   {
+    $this->authorize('update', [self::class, $article]);
     $request->validate([
       'datePublic' => 'required',
       'title' => 'required',
@@ -92,6 +93,7 @@ class ArticleController extends Controller
   public function destroy(Article $article)
   {
     // return response()->json($article->delete(), 201);
+    $this->authorize('delete', [self::class, $article]);
     $article->delete();
     return redirect()->route('article.index');
     // return response()->json($article, 201);
